Crate problem035

Problem 35 (Circular primes)

Solution detail

First, note that circular primes, other than the single digit primes 2 and 5, must consist entirely of the digits 1, 3, 7 and 9, since all other primes end in these digits.

This means that we can simply consider all combinations of these digits, and for each one, check if the corresponding number is a circular prime. We do this prime checking using a sieve which holds enough pre-computed prime numbers to do primality checking for each number we might encounter.

Structs

CircularPrimeTree
The information that is held about the current state during the tree search.

CircularPrimeTreeStep
A description of a step that can be taken in the search tree.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the number of circular primes there are with at most the given number of digits. Remember to add on 2 for the primes 2 and 5 which are not considered otherwise.